# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MobiManager API
  description: Complete API documentation for MobiManager - a mobile shop management SaaS platform
  version: 1.0.0
  contact:
    name: MobiManager Support
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - cookieAuth: []
paths:
...
  /api/admin/shops/{shopId}:
    get:
      summary: Get Shop
      tags: [Shops]
      parameters:
        - name: shopId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Shop details
    put:
      summary: Update Shop
      tags: [Shops]
      parameters:
        - name: shopId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Shop updated
    delete:
      summary: Delete Shop
      tags: [Shops]
      parameters:
        - name: shopId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Shop deleted

  # Admin Sub-Admin Routes
...
  /api/admin/reports/repairs:
    get:
      summary: Get Repairs Report
      tags: [Reports]
      parameters:
        - name: period
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Repairs report
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_token